Conflict resolution: import lists and registry entries taken as the union of both sides;
merging_press.py and its tests keep the branch (BasePress-capable) version; the ScorerPress
path n_kept now uses kvpress.utils.compute_n_kept as upstream does.

+++ kvpress/presses/cur_press.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 1993-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 1993-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
 import math
@@ -35,7 +35,7 @@ class CURPress(ScorerPress):
 
         if self.use_random_leverage:
             r = 20
-            G = torch.randn(keys.shape[-1], r, device=keys.device) / math.sqrt(r)
+            G = torch.randn(keys.shape[-1], r, device=keys.device, dtype=keys.dtype) / math.sqrt(r)
             keys = keys @ G
             values = values @ G
